fix: Prevent malicious websites from connecting to http / websocket server (#1157)
* fix: Prevent malicious websites from connecting to http / websocket server

Browsers allow any website to connect to websockets on `127.0.0.1` and,
therefore, to spy on users of tinymist. Disallow this by checking the
`Origin` header. Note: This does not protect against malicious users
that share the same `127.0.0.1` as us (e.g. multi-user systems where
the users don't trust each other). That requires additional changes
that may be added in the future.

* Add VSCode exception

* Also prevent malicious connections to trace/profiling http server

Allow VSCode only for now.