feat: scroll preview with hjkl (#2111)
Closes #2107 

Sorry for not getting confirmation on the issue first, but I had some
free time and thought it would be easy to implement and, indeed, it is
just a few lines of code. I'll attach a video of how it looks.


https://github.com/user-attachments/assets/60e6f7ef-cdff-4f85-a1e8-385058f87966

For the small nudges I think the instant jump is better, even though it
looks kind of janky. The problem with the smooth scroll is that you
can't press it repeatedly and I think that you use `jk` for small
adjustments where it's clear what is being scrolled. However, for the
big hl jumps I think the smooth scroll is better to not disorient you.

Happy to change anything, of course :)

---------

Co-authored-by: Myriad-Dreamin <camiyoru@gmail.com>