feat: typst-preview: add hotkey to toggle light/dark theme
This commit adds a hotkey "t" (abbreviated from theme) to the
typst-preview frontend which allows for quickly changing the rendered
theme. This is useful as it allows people working in dark mode
to quickly check how the final rendered PDF (which is in light mode)
would look without having to change their browser/system theme back and
forth.